[WIP] Writer: Add 'relative_x', 'relative_y' for all layout nodes
Currently, this is broken, I don't know what exactly is going on, but we
appear to be creating new regions even when the current region can fit
the page.

This doesn't help much without any means to calculate the actual width
of a layout node, but it should prevent the text from being drawn all
over the place.