Writer: Resolve some issues with reusing paragraphs
There was one typo where I used 'get_max_width' instead of
'get_max_inner_width'.  On top of that, I was checking if some floating
point numbers were exactly equal.

My code seems to assume that the absolute position is known which might
not be the case.  I have not addressed that problem yet.