Writer: Resolve two more issues with reusing layout nodes
First, when we reuse layout nodes we do not need to assume that we are
in the final phase.  The absolute position is not required.

Second, when clearing the cached absolute position and size values, I
forgot to update the phase, this is fixed now.

It seems that we crash again, possibly because I fixed these issues
improperly.  More investigation is required.