Writer: Update the model to allow placing cursor in model nodes
After some thought, I opted to add an optional offset to text chunk
model nodes.  The alternative would be to split model nodes, I do not
know what is better.

This mostly seems to work, but the cursor is not rendered, the cause
appears to be that the 'CursorLayoutNode.paint_decorations' method isn't
called.