Writer: Don't keep reference to parent style in 'ModelStyle'
When we merge paragraphs, this messes up the reference to the parent
style, that appears to be one of the issues we run into.  Instead, the
cascade should occur in the converter using the new 'ModelStyleCascade'
helper.

That is not yet implemented, we crash on startup.  The next step is to
use the new 'ModelStyleCascade' class in the converter module.