Writer: Remove previous converter logic and prepare to reimplement it
The previous conversion logic was a huge mess.  I am relatively sure
that I can do better than that.  For now I just deleted everything and
started writing an outline for what I am trying to do.

Initially, I wanted to do it very differently, however, I've now
reconsidered and the new implementation will be very similar to the
previous one.

Currently, we are presented with a blank page.  The next step would be
to implement all the logic for placing paragraphs again.