Debug: Pending paragraphs should always fit on the pending page
Actually, now that I write this, I already regret it.  This introduces
another edge case that is difficult to fix.  If we create a new pending
paragraph we must create a new pending page for that paragraph.  This
could cause us to add an unnecessary page in the end.

I think the problem is that I have a pending paragraph that doesn't
contain anything.  The invariant that I always have a pending paragraph
makes no sense.  That should be what I fix.

Currently, I crash for some reason and I haven't fully diagnosed why.  I
think I should roll back and then remove the invariant that there is
always a pending paragraph.  Maybe I can even move that into the
function instead of having this rather "global" variable.